Dev3.0

+++ NB2/首页/匹配中心/MarryConrtolViewController.h
@@ -9,5 +9,5 @@
 #import <UIKit/UIKit.h>
 
 @interface MarryConrtolViewController : UIViewController<TopViewDelegate>
-@property(nonatomic,assign) int indexsign;
+@property(nonatomic,assign) NSInteger indexsign;
 @end
